// packages/util/src/declaration-parser/types.ts
export interface ArrayPattern {
  type: 'ArrayPattern';
  elements: (DestructuringPattern | null)[];
}
export interface ArrowFunctionExpression {
  type: 'ArrowFunctionExpression';
  generator: boolean;
  id: null;
  params: Parameter[];
  // body: BlockStatement | Expression;
  async: boolean;
  // expression: boolean;
}
export interface AssignmentPattern {
  type: 'AssignmentPattern';
  left: BindingName;
  // right: Expression;
}
export type BindingName = BindingPattern | Identifier;
export type BindingPattern = ArrayPattern | ObjectPattern;
interface ClassDeclarationBase {
  id: Identifier | null;
  // body: ClassBody;
  // superClass: LeftHandSideExpression;
}
export interface ClassDeclaration extends ClassDeclarationBase {
  type: 'ClassDeclaration';
}
export interface ClassExpression extends ClassDeclarationBase {
  type: 'ClassExpression';
}
export type DestructuringPattern =
  | ArrayPattern
  | AssignmentPattern
  | Identifier
  // | MemberExpression
  | ObjectPattern
  | RestElement;
// ...
export interface Identifier {
  type: 'Identifier';
  name: string;
}
// ...
export interface ObjectPattern {
  type: 'ObjectPattern';
  properties: (Property | RestElement)[];
}
export type Parameter =
  | ArrayPattern
  | AssignmentPattern
  | Identifier
  | ObjectPattern
  | RestElement;
export type Property =
  // | PropertyComputedName
  PropertyNonComputedName;
// ...
export interface RestElement {
  type: 'RestElement';
  argument: DestructuringPattern;
}